requirements.txt
streamlit>=1.35.0
google-generativeai>=0.8.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
pandas>=2.2.0
streamlit-webrtc>=0.47.0
SpeechRecognition>=3.10.0
fpdf2>=2.7.0
